ofproto-dpif-upcall: Filter translated mask using flow
When revalidating a flow convert xout.wc.masks, the mask calculated by
translation, to a key that relates to the flow present in the datapath.  Only
mask elements that relate to the flow will be included in the resulting key.
This has the effect of filtering the mask based on 'flow'.

In particular this filters out mpls_lse mask bits set on non-MPLS flows.

This avoids flows being premeturely evicted from the datapath
because their calculated mask is too specific.

Signed-off-by: Simon Horman <horms@verge.net.au>