ofproto-dpif: Execute with helper if recirc actions are present
In the case where packet is being processed by
ofproto_dpif_execute_actions*() it may have an in_port
that does not exist in the datapath: e.g. a packet from a packet_out message
with the CONTROLLER as the in_port.

In such cases it seems to make little sense to perform execution
in the datapath if a recirc action is present as it will cause
an upcall with an in_port that doesn't exist. Indeed, in this case
packets are dropped.

Resolve this problem by executing packets in ovs-vswtichd if
they handled by ofproto_dpif_execute_actions*() and translation
adds a recirc actions.

Signed-off-by: Simon Horman <horms@verge.net.au>