dpif: Allow hash actions in dpif_execute_helper_cb()
Rather then prohibiting hash actions dpif_execute_helper_cb()
execute them by passing them on to aux->dpif->dpif_class->execute()
and save the updated packet metadata afterwards. It is assumed
that if a hash is calculated it will be sored in the dp_hash field
of the metadata.

This is in preparation for allowing execution of recirc actions
in ovs-vswitchd.

Signed-off-by: Simon Horman <horms@verge.net.au>