datapath: Free skb(s) on recirculation error
It is my understanding that on error execute_recirc() does not free the
skb passed to it.

Assuming this is true then on error skb should always be freed
if an error occurs in execute_recirc().

And if recirc_skb differs from skb, because it is a clone of skb,
then it should also be freed.

Signed-off-by: Simon Horman <horms@verge.net.au>