ofp-util: Use enum ofp14_flow_update_event in struct ofputil_flow_update
Switch to using enum ofp14_flow_update_event instead of
enum nx_flow_update_event in struct ofputil_flow_update as
the former can encode all values defined for the latter but
the reverse is not true.

Also use switch rather than if when testing event values
to give a little bit of extra type checking.

This is in preparation for supporting OpenFlow 1.4 flow monitor replies.

Signed-off-by: Simon Horman <horms@verge.net.au>