ofp-parse: Correct sense of NO_ABBREV flag in flow monitor request parser
Unlike other flags when '!own' is parsed it should set
rather than clear the OFPFMF14_NO_ABBREV flag.

This is because the default (no '!own') corresponds to
the behaviour when NO_ABBREV is clear (i.e. do abbreviate).
And the alternate behaviour ('!own' present) corresponds to
the behaviour when NO_ABBREV is set (i.e. don't abbreviate).

A subsequent patch will update ovs-vswtichd so that it
uses the NO_ABBREV flag: Currently it abbreviates regardless
of if NO_ABBREV is set or not.

Signed-off-by: Simon Horman <horms@verge.net.au>