ofp-parse: Correctly update bucket lists if they are empty
list_moved() appears to only work reliably if its argument is a non-empty
list. This is because list_moved dereferences the next and prev fields
of the list, which for an empty list are the address of the list itself,
and list_moved() is intended to be called when the address of the list has
changed.

This patch updates the only user of list_moved() that I was able to find
my temporarily setting the next field to NULL if it is empty and then
re-initialising the list rather than calling list_moved() after
reallocation has occurred.

An test update is also provided to exercise this change. Without
the code change the updated test cases a segmentation when the
buckets parameter of ofputil_put_ofp11_bucket() is dereferenced.

Signed-off-by: Simon Horman <simon.horman@netronome.com>

---

* Although somewhat cure the approach of setting the next field to NULL
  seems far less dangerous than trying to update the list infrastructure
  to handle this case.

  I also toyed with using a temporary bitmap to track empty buckets.
  That approach worked but seemed to be more overhead and no cleaner
  than the one taken by this patch.