userspace: add non-tap (l3) support to GRE vports
Add support for layer 3 GRE vports (non-tap aka non-VTEP).

This makes use of a separate vport type for GRE, rather than a new mode for
the existing (tap/VTEP) GRE vports as this fits more naturally with the
kernel where implementation of GRE and thus implementation of this feature
there.

In order to differentiate packets for two different types of GRE vports a
new flow key attribute, OVS_KEY_ATTR_NEXT_BASE_LAYER, is used.  It is
intended that this attribute is only used in userspace as there appears to
be no need for it to be used in the kernel datapath.

It is envisaged that this attribute may be used for other non-UDP
encapsulation protocols that support both layer3 and layer2 inner-packets.
While for UDP encapsulation the UDP port can be used for differentiation
without the need for this new attribute.

One alternative approach to this new attribute, which I have not
investigated in detail, would be to use a second classifier in tnl-ports.c
for non-UDP layer3 tunnels; leaving the existing classifier for all other
tunnels.

Signed-off-by: Simon Horman <simon.horman@netronome.com>

---
v9
* New patch