userspace: add non-tap (l3) support to GRE vports
Add support for layer 3 GRE vports (non-tap aka non-VTEP).

This makes use of a vport mode configuration for the existing (tap/VTEP)
GRE vports.

In order to differentiate packets for two different types of GRE vports a
new flow key attribute, OVS_KEY_ATTR_NEXT_BASE_LAYER, is used.  It is
intended that this attribute is only used in userspace as there appears to
be no need for it to be used in the kernel datapath.

It is envisaged that this attribute may be used for other encapsulation
protocols that support both layer3 and layer2 inner-packets.

Signed-off-by: Simon Horman <simon.horman@netronome.com>

---
v11
* Update for new kernel implementation that uses ARPHRD_NONE devices
* Omit match_base_layer from struct tun_port_in, it is not necessary any more
* Correct OVS_KEY_ATTR_NEXT_BASE_LAYER handling in odp_flow_key_from_flow__()
  and parse_l2_5_onward() so that the encoding and decoding of the
  next_base_layer field of a flow key is consistent both with itself on
  encode and decode and with other fields of the flow key.
* Guard OVS_KEY_ATTR_TUNNEL_INFO with #ifndef __KERNEL__ in
  openvswitch.h and note that it is only used by the user-space datapath
* Consistently use tabs for indentation
* Correct spelling and capitalisation of documentation
* Append "(layer3)" to when showing layer3 tunnel ports as
  they may share the same ODP port as a non-layer3 tunnel port.
  This seems simpler than v10 which included an O(n**2) filter on
  showing duplicate ports.

v10
* Use a mode for layer3 ports rather than a new port type
* Update BUILD_BUG_ON() call in ovs_key_attr_size()
* Don't update tnl_port_map_lookup() to always match
  on next_base_layer: the implementation didn't actually
  do that and thus was a lot of code change for no behavioural
  change.

v9
* New patch

fix

Signed-off-by: Simon Horman <simon.horman@netronome.com>