userspace: extend layer 3 support to cover non-IP packets
Extend support for layer 3 packets to cover non-IP packets.

This removes the assumption that the first octet of a layer 3 packet
indicates the IP protocol version - true for IP (v4 and v6), but not
for necessarily for other protocols.

The key motivation for this is to allow forwarding of MPLS packets which
are technically layer 2.5 rather than 3 but the distinction seems unimportant
here.

This relies on datapaths setting OVS_KEY_ATTR_PACKET_ETHERTYPE to
the ethernet type corresponding to the protocol of layer 3 packets
on a flow miss.

Signed-off-by: Simon Horman <simon.horman@netronome.com>
Acked-by: Ben Pfaff <blp@ovn.org>

---
v11
* Added Ack from Ben Pfaff

v10
* Delete now bogus comment about protocol assumptions for L3 packets

v9
* New patch