wip: packets: set packet_ethtype in metadata of l3 packets
Without this change setting the ethernet type of a packet
in the user-space datapath, e.g. when performing a pop_mpls action,
nothing will occur for l3 packets due to the absence of an ethernet header.
The correct remedy appears to be to set the ethernet type in metadata.

Without this change translation of actions after recirculation,
for example due to pop mpls, may not be done correctly resulting
in incorrect datapath flows and thus incorrect packet forwarding
or dropping behaviour.

Signed-off-by: Simon Horman <simon.horman@netronome.com>