userspace: correct handing of next_base_layer in flow key
Correct odp_flow_key_from_flow__() and parse_l2_5_onward()
so that the encoding and decoding of the next_base_layer field
of a flow key is consistent both with itself on encode and decode
and with other fields of the flow key.

This corrects flow revalidation and allows flows to be modified
e.g. to reflect updated state of members of a bond detected by LCAP.

Fixes: 582efbd603ac ("userspace: add non-tap (l3) support to GRE vports")
Signed-off-by: Simon Horman <horms@verge.net.au>