userspace: Enable IP checksum offloading by default.
The netdev receiving packets is supposed to provide the flags
indicating if the IP checksum was verified and it is GOOD or BAD,
otherwise the stack will check when appropriate by software.

If the packet comes with good checksum, then postpone the
checksum calculation to the egress device if needed.

When encapsulate a packet with that flag, set the checksum
of the inner IP header since that is not yet supported.

Calculate the IP checksum when the packet is going to be sent over
a device that doesn't support the feature.

Linux devices don't support IP checksum offload alone, so the
support is not enabled.

Signed-off-by: Flavio Leitner <fbl@sysclose.org>
Co-authored-by: Mike Pattrick <mkp@redhat.com>
Signed-off-by: Mike Pattrick <mkp@redhat.com>
--
 Since v9:
  - Removed duplicative field tx_ip_csum_offload from netdev-dpdk.c
  - Left rx_csum_offload field as it is not duplicative
  - Moved system-userspace-offload.at tests to dpif-netdev.at
  - Various visual changes
  - Extended miniflow_extract changes into avx512 code
 Since v10:
  - avx512 checksum length corrected