vswitch.xml: Use member wording for bonds.
Since the patch-set that included [1] there has been a policy of using
the term member for bonds, LACP, and bundle contexts.  This is
consistent with the more recently adopted policy of using the inclusive
naming word list v1 [2, 3].

This patch addresses two instances where the term member should be used
in vswitch.xml. It does not address instances of alternative wording
that require code updates, which can addressed as follow-up activity.

[1] 91fc374a9c5a ("Eliminate use of term "slave" in bond, LACP, and bundle contexts.")
[2] df5e5cf4318a ("Documentation: Add section on inclusive language.")
[3] https://inclusivenaming.org/word-lists/

Signed-off-by: Simon Horman <horms@ovn.org>