wip: util: Add ovs_hard_stop
Recently OVS adopted a policy of using the inclusive naming word list v1
[1, 2].

This patch addresses the use of the term abort within the scope of
direct calls to the library function abort(). It does so by providing
a wrapper ovs_hard_stop(), thus reducing the incidence of the term
abort in the source code.

[1] df5e5cf ("Documentation: Add section on inclusive language.")
[2] https://inclusivenaming.org/word-lists/

Signed-off-by: Simon Horman <horms@ovn.org>