wip: util: Rename ovs_abort() as ovs_force_stop()
Recently OVS adopted a policy of using the inclusive naming word list v1
[1, 2].

This patch addresses the use of the term abort within the scope of the
functions ovs_abort(), and ovs_force_abort_valist(). This is done by
renaming them as ovs_force_stop(), and ovs_force_stop_valist().

[1] df5e5cf ("Documentation: Add section on inclusive language.")
[2] https://inclusivenaming.org/word-lists/

Signed-off-by: Simon Horman <horms@ovn.org>