wip: util: Rename vlog_abort() as vlog_force_stop()
Recently OVS adopted a policy of using the inclusive naming word list v1
[1, 2].

This patch addresses the use of the term abort within the scope of the
functions vlog_abort() and vlog_abort_valist(), and the macro
VLOG_ABORT(). This is done by renaming them as vlog_force_stop(),
vlog_force_stop_valist(), and VLOG_FORCE_QUIT().

As the renamed functions and macro are exported as an API to third-party
code, legacy implementations using the old names are provided for
compatibility. As a follow-up scheme to remove these over time may be
developed.

[1] df5e5cf ("Documentation: Add section on inclusive language.")
[2] https://inclusivenaming.org/word-lists/

Signed-off-by: Simon Horman <horms@ovn.org>