wip: json: Rename json_parser_abort() as json_parser_hard_stop()
Recently OVS adopted a policy of using the inclusive naming word list v1
[1, 2].

This patch addresses the use of the term abort within the scope of the
function json_parser_abort() by renaming it as json_parser_hard_stop().

As the renamed function are exported as an API to third-party code,
a legacy implementation is provided using the old name for compatibility.
As a follow-up scheme to remove these over time may be developed.

[1] df5e5cf ("Documentation: Add section on inclusive language.")
[2] https://inclusivenaming.org/word-lists/

Signed-off-by: Simon Horman <horms@ovn.org>