ci: Run system tests in a separate namespace.
GitHub runners use 10.1.0.0/16 network as their base network for eth0
interface.  That is causing random system test failures when unexpected
conntrack entries for this network are present, because our system
tests are mainly using 10.1.1.0/24 subnet for their test networks.

Run system tests in their own network namespace to avoid any unwanted
interference.

Ideally, we would run every single test in its own namespace, but that
is not a trivial change and will likely be hard to backport.  Still
worth investigating in the future.

Note: Layer3 tunnel tests with Bareudp ports rely on loopback to work,
but lo interface is down by default in new namespaces.  So, bringing
it up.  These tests are skipped in Ubuntu 22.04, because it doesn't
have bareudp support, but it's better to have the change anyway, so it
doesn't bite us in the future while upgrading the base image.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Simon Horman <horms@ovn.org>