wip: ci: Use the term hard stop in netdev statistics
Recently OVS adopted a policy of using the inclusive naming word list v1
[1, 2].

This patch addresses the use of the term abort within the scope of
netdev statistics. It does so by using the field tx_aborted_errors
as tx_hard_stop_errors.

The primary target for this change is struct netdev_stats.

struct rtnl_link_stats64 is also updated as although this
is a Linux kernel structure, OVS unconditionally provides
a replacement definition.

[1] df5e5cf ("Documentation: Add section on inclusive language.")
[2] https://inclusivenaming.org/word-lists/

Signed-off-by: Simon Horman <horms@ovn.org>