wip: ovsdb-idl: Rename aborted status as hard stop
Recently OVS adopted a policy of using the inclusive naming word list v1
[1, 2].

This patch addresses the use of the term abort within the status
returned, as a string, when ovsdb idl performs a hard stop.

This covers:
* The C implementation
* The Python implementation
* The test suite

A follow-up patch will further replace the use of the term
abort in the C and Python implementations.

[1] df5e5cf ("Documentation: Add section on inclusive language.")
[2] https://inclusivenaming.org/word-lists/

Signed-off-by: Simon Horman <horms@ovn.org>