doc: Require Sphinx version 4.0
Require Sphinx version 4.0, Sphinx 4.0.0 was released on 9th May 2021 [1].

The motivation for this to allow use of the root_doc directive in place
of master_doc, which will be the subject of a subsequent patch.

I have made some effort to update dependencies accordingly.
A  exception being .cirrus.yml, where I am entirely unsure
of the applicability of versioned dependencies.

This patch also drops support for using sphinx-build-2 because Sphinx
4.0 doesn't support Python 2.

Some implications for Linux distributions are:

* Debian: python3-sphinx >= 4.0 has been available since Bookworm
* Fedora: python-sphinx >= 4.0 is available since at least Fedora 38
* RHEL:   Sphinx 3.0 is not available as of RHEL 9
* Ubuntu: python3-sphinx >= 4.0 has been available since Jammy (22.04)

[1] https://pypi.org/project/Sphinx/#history

Signed-off-by: Simon Horman <horms@ovn.org>