appveyor: Prepare for rename of primary development branch
Recently OVS adopted a policy of using the inclusive naming word list v1
[1, 2].  And in keeping with this policy it is intended to rename the
primary development branch from master to main [3].

In order to help facilitate this change allow Appveyor to run
on the main as well as master branch. It is intended that
master branch will be removed from appveyor.yml after the primary branch
has been renamed.

Also, in keeping with the inclusive naming policy, update other
locally controlled instances of 'master' to 'main' in this file.
(There is an external URL which includes 'master' which cannot be
 updated without causing breakage.)

[1] df5e5cf ("Documentation: Add section on inclusive language.")
[2] https://inclusivenaming.org/word-lists/
[3] https://mail.openvswitch.org/pipermail/ovs-dev/2024-March/412686.html

Signed-off-by: Simon Horman <horms@ovn.org>