tests: add nft accept support.
Certain Linux distributions, like CentOS, have default iptable rules
to reject input traffic from bridges such as br-underlay.

To address this, IPTABLES_ACCEPT adds an iptables rule to always accept
the traffic.

As part of an effort to use nft in place of iptables in the testsuite,
implement NFT_ACCEPT, an nft version of IPTABLES_ACCEPT. As the
condition where IPTABLES_ACCEPT implies the existence of an INPUT chain,
only instantiate an nft rule in that chain if it already exists.

Also provide a wrapper, XT_ACCEPT, which will call NFT_ACCEPT if
nft is available, and IPTABLES_ACCEPT otherwise

And provide OVS_CHECK_XT, which can be used to check if the
prerequisites for running XT_ACCEPT are present, and skips the current
test otherwise.

Update the one test where IPTABLES_ACCEPT is used so that it
now uses XT_ACCEPT and OVS_CHECK_XT.

Signed-off-by: Simon Horman <horms@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>