NEWS: Deprecate running user-space with OOT Kernel module.
It is some time now since the OOT kernel module was removed. It is yet
longer since it was deprecated. And it is no longer present in any
supported versions.

However, running user-space with the OOT kernel module, from unsupported
releases, is still supported.

This notice deprecates that support as of the v3.7 release.
With the intention to remove it in the v3.8 release.

As discussed at OVS+OVN Conference 2025.

Acked-by: Eelco Chaudron <echaudro@redhat.com>
Signed-off-by: Simon Horman <horms@ovn.org>

--
v2:
- Minor edit to NEWS text
- Add Acked-by tag from Eelco
- Incorporated suggestions from Ilya
  + Use NEWS rather than News as patch prefix: it's the file's name
  + Add run-time deprecation warning message
  + Document that the last supported release with the OOT module was
    v2.17 rather than referring to the module's removal in v3.0.
  + Typo: use-space -> user-space
- Link to v1: https://mail.openvswitch.org/pipermail/ovs-dev/2025-December/428402.html