dpif-netlink-rtnl: Tighten probe for OOT Kernel tunnels.
It is common to run make check as an unprivileged user.
And in that case calling dpif_netlink_rtnl_create() to add a tunnel
fails due to insufficient privileges. Which is assumed
to mean that OOT Kernel tunnels are used.

Reverse this assumption and assume they are not being used.
An assumption in making this change is that the distinction
is unimportant in terms of run-time tunnel usage as as
make check doesn't use kernel tunnels.

However, a follow-up patch will add a deprecation notice, which prints
in the case that OOT Kernel tunnels are used.  It does not seem useful
to print in the case of make check.  And doing so breaks existing tests
run by that target.

Signed-off-by: Simon Horman <horms@ovn.org>
--
v2: New patch

This patch works on the principle that small, subtle changes,
are accompanied by explanations and comments significantly larger
than the code change.